http / ws별파싱하여 action call -> response

+++ src/lib/redis/index.ts
@@ -49,8 +49,10 @@ class Redis {
    * @returns {void}
    */
   remove = (key: string): void => {
-    this.client.del(key);
-    console.log(`============> redis remove ${key}`);
+    if (!_.isEmpty(key)) {
+      this.client.del(key);
+      console.log(`============> redis remove ${key}`);
+    }
   };
 
   /**
